requirements.txt
slack-sdk==3.21.3
slack-bolt==1.18.0
pyautogen==0.2.3
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
PyGithub==1.58.2
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
slack-bolt==1.18.0
supabase==2.3.5
pydantic==2.6.1 